feat: Move the RawDocument type to its own module
